+++ NEWS.md
@@ -1,5 +1,7 @@
 # saperlipopette (development version)
 
+* New `exo_revparse()` exercise to practice parsing ancestry references with `git rev-parse`.
+
 * New `exo_worktree()` exercise to practice `git worktree`.
 
 * Add instructions with "switch" and "restore" alongside "checkout" equivalents (#30).